pyproject.toml
[project]
name = "rbc-pinn-surrogate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { email = "tmarkmann@techfak.uni-bielefeld.de" }
]
requires-python = ">=3.11"
dependencies = [
    "ffmpeg>=1.4",
    "h5py>=3.14.0",
    "hydra-colorlog>=1.2.0",
    "hydra-core>=1.3.2",
    "lightning[extra]>=2.5.5",
    "neuraloperator>=1.0.2",
    "pandas>=2.3.0",
    "rbc-gym",
    "rich>=14.0.0",
    "seaborn>=0.13.2",
    "torch==2.7.1",
    "torch-harmonics>=0.7.3",
    "torchmetrics>=1.7.3",
    "torchvision>=0.21.0",
    "wandb==0.20.1",
]

[project.scripts]
rbc-pinn-surrogate = "rbc_pinn_surrogate:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ruff>=0.11.6",
]

[tool.uv.sources]
rbc-gym = { git = "https://github.com/HammerLabML/RBC-Gym" }
